rename BS content card components
