+++ components/vault/detailsSection/ContentCardTargetColRatioAfterBuy.tsx
@@ -4,25 +4,26 @@ import {
   ContentCardProps,
   DetailsSectionContentCard,
 } from 'components/DetailsSectionContentCard'
+import { maxUint256 } from 'features/automation/common/basicBSTriggerData'
 import { formatAmount, formatPercent } from 'helpers/formatters/format'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 
-interface ContentCardTargetSellColRatioProps {
+interface ContentCardTargetColRatioAfterBuyProps {
   targetColRatio?: BigNumber
   afterTargetColRatio?: BigNumber
-  threshold?: BigNumber
+  threshold: BigNumber
   changeVariant?: ChangeVariantType
-  token: String
+  token: string
 }
 
-export function ContentCardTargetSellColRatio({
+export function ContentCardTargetColRatioAfterBuy({
   targetColRatio,
   afterTargetColRatio,
   threshold,
   changeVariant,
   token,
-}: ContentCardTargetSellColRatioProps) {
+}: ContentCardTargetColRatioAfterBuyProps) {
   const { t } = useTranslation()
 
   const formatted = {
@@ -38,11 +39,13 @@ export function ContentCardTargetSellColRatio({
         precision: 2,
         roundMode: BigNumber.ROUND_DOWN,
       }),
-    threshold: threshold && `$${formatAmount(threshold, 'USD')}`,
+    threshold: threshold.isEqualTo(maxUint256)
+      ? t('unlimited')
+      : `$${formatAmount(threshold, 'USD')}`,
   }
 
   const contentCardSettings: ContentCardProps = {
-    title: t('auto-sell.target-col-ratio-after-selling'),
+    title: t('auto-buy.target-col-ratio-after-buying'),
     value: formatted.targetColRatio,
   }
 
@@ -52,13 +55,13 @@ export function ContentCardTargetSellColRatio({
       variant: changeVariant,
     }
   if (threshold)
-    contentCardSettings.footnote = t('auto-sell.continual-sell-threshold', {
+    contentCardSettings.footnote = t('auto-buy.continual-buy-threshold', {
       amount: formatted.threshold,
       token,
     })
 
-  if (!threshold || threshold?.isZero())
-    contentCardSettings.footnote = t('auto-sell.continual-sell-no-threshold')
+  if (!threshold || threshold.isEqualTo(maxUint256) || threshold.isZero())
+    contentCardSettings.footnote = t('auto-buy.continual-buy-no-threshold')
 
   return <DetailsSectionContentCard {...contentCardSettings} />
 }
